[REF] core: separated odoo.orm.environments and decorators
New packages for decorators and environments.
