+++ odoo/api/__init__.py
@@ -2,8 +2,7 @@
 # Exports features of the ORM to developers.
 # This is a `__init__.py` file to avoid merge conflicts on `odoo/api.py`.
 from odoo.orm.identifiers import NewId
-from odoo.orm.api import (
-    Environment,
+from odoo.orm.decorators import (
     autovacuum,
     call_kw,
     constrains,
@@ -16,5 +15,6 @@ from odoo.orm.api import (
     readonly,
     returns,
 )
+from odoo.orm.environments import Environment
 
 from odoo.orm.types import ContextType, DomainType, IdType, Self, ValuesType
